move call logic of acir gen into its own module
